backend/pyproject.toml
[project]
name = "dzencode-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.3.0",
    "celery>=5.5.3",
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.1",
    "django>=5.2.8",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.9.0",
    "django-recaptcha>=4.1.0",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "strawberry-graphql-django>=0.73.1",
    "uvicorn[standard]>=0.38.0",
]
